 \docType{data}
    \name{sboner}
    \alias{sboner}
    \title{The Sboner Cohort}
    \format{a \code{ExpressionSet} instance, 1 row per Affymetrix microarray probe.}
    \source{
      GSE16560
    }
    \description{
      Object imported using GEOquery and curated to remove variables introduced by GEO
    }
    \author{
Mark Dunning
    }
    \usage{sboner}
    \keyword{datasets}